Lagt til gui i display.py og api i api.py. Fullført opprett_kunde funksjon
